+++ .github/workflows/linux.yml
@@ -134,6 +134,7 @@ jobs:
 
       - name: Release
         uses: softprops/action-gh-release@v2
+        if: ${{ github.ref == 'refs/heads/master' }}
         with:
           tag_name: "Nightly"
           files: |
